feat: add support for providing custom values in the duration filter
